Refactor Parser to remove wildcard import
