+++ src/main/java/friday/Friday.java
@@ -1,9 +1,9 @@
 package friday;
 
 import friday.command.Command;
+import friday.command.Parser;
 import friday.task.TaskList;
 import friday.util.FridayException;
-import friday.util.Parser;
 import friday.util.Storage;
 import friday.util.Ui;
